tools: isomaker: refactor as a lib instead of main package
